fixed build and tests
